backend/requirements.txt
annotated-types==0.7.0
anyio==4.13.0
async-timeout==5.0.1
bcrypt==3.2.2
cffi==2.0.0
click==8.3.3
colorama==0.4.6
dnspython==2.8.0
email-validator==2.3.0
exceptiongroup==1.3.1
fastapi==0.110.0
greenlet==3.4.0
h11==0.16.0
idna==3.13
passlib==1.7.4
psycopg2-binary==2.9.9
pycparser==3.0
pydantic==2.6.4
pydantic_core==2.16.3
python-dotenv==1.0.1
redis==7.4.0
SQLAlchemy==2.0.29
starlette==0.36.3
typing_extensions==4.15.0
uvicorn==0.29.0